requirements.txt
pypdftk
pdftotext
PyPDF2
easygui
Pillow
fpdf